Refactor file structure and update import paths for controllers and database connection
