+++ src/middlewares/verifyTokens.js
@@ -13,7 +13,7 @@ export const verifyTokens = (req, res, next) => {
             return res.status(401).json({ message: 'Unauthorized: Required tokens missing!' });
         }
         
-        const decodedUser = jwt.verify(userToken, `process.env.JWT_KEY`);
+        const decodedUser = jwt.verify(userToken, process.env.JWT_KEY);
         req.user = decodedUser;
         // console.log(decodedUser);
         next();
